:recycle: separate service file
